IdeaBoard - handle threads

+++ ideaboard/commands/adminbase.py
@@ -82,18 +82,22 @@ class AdminBase(MixinMeta):
             return
 
         content = message.embeds[0].description
-        thread: discord.Thread = await ctx.guild.fetch_channel(suggestion.thread_id)
-        if thread:
-            if conf.delete_threads:
-                with suppress(discord.HTTPException):
-                    await thread.delete()
-            else:
-                # Close and lock the thread
-                newname = thread.name + _(" [Approved]")
-                embed = discord.Embed(color=discord.Color.green(), description=content, title=_("Approved Suggestion"))
-                with suppress(discord.HTTPException):
-                    await thread.send(embed=embed)
-                    await thread.edit(archived=True, locked=True, name=newname)
+        if suggestion.thread_id:
+            with suppress(discord.NotFound):
+                thread: discord.Thread = await ctx.guild.fetch_channel(suggestion.thread_id)
+                if thread:
+                    if conf.delete_threads:
+                        with suppress(discord.HTTPException):
+                            await thread.delete()
+                    else:
+                        # Close and lock the thread
+                        newname = thread.name + _(" [Approved]")
+                        embed = discord.Embed(
+                            color=discord.Color.green(), description=content, title=_("Approved Suggestion")
+                        )
+                        with suppress(discord.HTTPException):
+                            await thread.send(embed=embed)
+                            await thread.edit(archived=True, locked=True, name=newname)
 
         embed = discord.Embed(color=discord.Color.green(), description=content, title=_("Approved Suggestion"))
         if author := ctx.guild.get_member(suggestion.author_id):
@@ -217,18 +221,22 @@ class AdminBase(MixinMeta):
             return
 
         content = message.embeds[0].description
-        thread: discord.Thread = await ctx.guild.fetch_channel(suggestion.thread_id)
-        if thread:
-            if conf.delete_threads:
-                with suppress(discord.HTTPException):
-                    await thread.delete()
-            else:
-                # Close and lock the thread
-                newname = thread.name + _(" [Rejected]")
-                embed = discord.Embed(color=discord.Color.red(), description=content, title=_("Rejected Suggestion"))
-                with suppress(discord.HTTPException):
-                    await thread.send(embed=embed)
-                    await thread.edit(archived=True, locked=True, name=newname)
+        if suggestion.thread_id:
+            with suppress(discord.NotFound):
+                thread: discord.Thread = await ctx.guild.fetch_channel(suggestion.thread_id)
+                if thread:
+                    if conf.delete_threads:
+                        with suppress(discord.HTTPException):
+                            await thread.delete()
+                    else:
+                        # Close and lock the thread
+                        newname = thread.name + _(" [Rejected]")
+                        embed = discord.Embed(
+                            color=discord.Color.red(), description=content, title=_("Rejected Suggestion")
+                        )
+                        with suppress(discord.HTTPException):
+                            await thread.send(embed=embed)
+                            await thread.edit(archived=True, locked=True, name=newname)
 
         embed = discord.Embed(color=discord.Color.red(), description=content, title=_("Rejected Suggestion"))
         if conf.anonymous and not conf.reveal:
